fix(tests): fold the denial tests into the mirror file (PS-204)
The audit requires a test file to mirror its src module 1:1.
test__host_exec_denial_names_the_case.py implied a src module of that
name; the module under test is _host_exec.py, which already has
test__host_exec.py. So the tests belong there, not in a second file.

They also reuse that file's existing _FakeRequest and _run helpers
instead of carrying duplicates, which is the better outcome regardless
of the rule.

31 passed in the merged file.
